runs without changing code each time
This py file uses a count.txt that holds the amount of times the program has been run. By setting the value in count.txt to 0 or deleting count.txt, you can create a data base by running the program 26 times. I also tried to write out comments and explanations for my functions. I also created a function that accesses the current database for dates from the Baseball table for dates for the Weather table so we do not need a global dictionary anymore.